+++ tests/integration/strategy_tests.cpp
@@ -39,7 +39,7 @@
 #include <omp.h>
 #endif
 
-// static short alignment = 16;
+static short alignment = 16;
 
 static int unique_pool_name = 0;
 static int initial_min_size = 1024;
@@ -128,7 +128,7 @@ TEST_P(StrategyTest, Duplicate)
 INSTANTIATE_TEST_CASE_P(Allocations, StrategyTest, ::testing::ValuesIn(AllocationDevices),);
 
 #if defined(UMPIRE_ENABLE_DEVICE)
-TEST(SimpoolStrategy, Device)
+TEST(Strategy, Device)
 {
   auto& rm = umpire::ResourceManager::getInstance();
 
@@ -150,9 +150,9 @@ TEST(SimpoolStrategy, Device)
   }
 
   allocator = rm.makeAllocator<umpire::strategy::DynamicPool>(
-      "device_simpool", rm.getAllocator("DEVICE"));
+      "device_dyn_pool", rm.getAllocator("DEVICE"));
 
-  ASSERT_EQ(allocator.getName(), "device_simpool");
+  ASSERT_EQ(allocator.getName(), "device_dyn_pool");
 
   ASSERT_NO_THROW( { alloc = allocator.allocate(100); } );
   ASSERT_GE(allocator.getCurrentSize(), 100);
@@ -388,7 +388,7 @@ TEST(ReleaseTest, Works)
   auto& rm = umpire::ResourceManager::getInstance();
 
   auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
-      "host_simpool_for_release", rm.getAllocator("HOST"), 64, 64);
+      "host_dyn_pool_for_release", rm.getAllocator("HOST"), 64, 64);
 
   void* ptr_one = alloc.allocate(62);
   void* ptr_two = alloc.allocate(1024);
@@ -404,7 +404,7 @@ TEST(DynamicPool, coalesce)
   auto& rm = umpire::ResourceManager::getInstance();
 
   auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
-      "host_pool_for_coalesce", rm.getAllocator("HOST"));
+      "host_dyn_pool_for_coalesce", rm.getAllocator("HOST"));
 
   auto strategy = alloc.getAllocationStrategy();
   auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
@@ -436,185 +436,185 @@ TEST(DynamicPool, coalesce)
   ASSERT_GE(dynamic_pool->getHighWatermark(), 62 + 1024);
 }
 
-// TEST(HeuristicTest, OutOfBounds)
-// {
-//   EXPECT_THROW({
-//     auto h = umpire::strategy::heuristic_percent_releasable(-1);
-//     UMPIRE_USE_VAR(h);
-//   }, umpire::util::Exception);
+TEST(HeuristicTest, OutOfBounds)
+{
+  EXPECT_THROW({
+    auto h = umpire::strategy::heuristic_percent_releasable(-1);
+    UMPIRE_USE_VAR(h);
+  }, umpire::util::Exception);
 
-//   EXPECT_THROW({
-//     auto h = umpire::strategy::heuristic_percent_releasable(101);
-//     UMPIRE_USE_VAR(h);
-//   }, umpire::util::Exception);
-// }
+  EXPECT_THROW({
+    auto h = umpire::strategy::heuristic_percent_releasable(101);
+    UMPIRE_USE_VAR(h);
+  }, umpire::util::Exception);
+}
 
-// TEST(HeuristicTest, EdgeCases_75)
-// {
-//   auto& rm = umpire::ResourceManager::getInstance();
+TEST(HeuristicTest, EdgeCases_75)
+{
+  auto& rm = umpire::ResourceManager::getInstance();
 
-//   auto h_fun = umpire::strategy::heuristic_percent_releasable(75);
+  auto h_fun = umpire::strategy::heuristic_percent_releasable(75);
 
-//   auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
-//       "host_dyn_pool_h_75", rm.getAllocator("HOST"),
-//       1024ul, 1024ul, alignment, h_fun);
+  auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
+      "host_dyn_pool_h_75", rm.getAllocator("HOST"),
+      1024ul, 1024ul, alignment, h_fun);
 
-//   auto strategy = alloc.getAllocationStrategy();
-//   auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
+  auto strategy = alloc.getAllocationStrategy();
+  auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
 
-//   if (tracker) {
-//     strategy = tracker->getAllocationStrategy();
-//   }
+  if (tracker) {
+    strategy = tracker->getAllocationStrategy();
+  }
 
-//   auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
+  auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
 
-//   ASSERT_NE(dynamic_pool, nullptr);
+  ASSERT_NE(dynamic_pool, nullptr);
 
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* a[4];
-//   for ( int i = 0; i < 4; ++i ) {
-//     ASSERT_NO_THROW({ a[i] = alloc.allocate(1024); });
-//     ASSERT_EQ(alloc.getActualSize(), (1024*(i+1)));
-//     ASSERT_EQ(dynamic_pool->getBlocksInPool(), (i+1));
-//     ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
-//   }
+  void* a[4];
+  for ( int i = 0; i < 4; ++i ) {
+    ASSERT_NO_THROW({ a[i] = alloc.allocate(1024); });
+    ASSERT_EQ(alloc.getActualSize(), (1024*(i+1)));
+    ASSERT_EQ(dynamic_pool->getBlocksInPool(), (i+1));
+    ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  }
 
-//   for ( int i = 0; i < 2; ++i ) {
-//     ASSERT_NO_THROW({ alloc.deallocate(a[i]); });
-//     ASSERT_EQ(dynamic_pool->getBlocksInPool(), 4);
-//   }
+  for ( int i = 0; i < 2; ++i ) {
+    ASSERT_NO_THROW({ alloc.deallocate(a[i]); });
+    ASSERT_EQ(dynamic_pool->getBlocksInPool(), 4);
+  }
 
-//   ASSERT_NO_THROW({ alloc.deallocate(a[2]); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
+  ASSERT_NO_THROW({ alloc.deallocate(a[2]); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(a[3]); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-// }
+  ASSERT_NO_THROW({ alloc.deallocate(a[3]); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+}
 
-// TEST(HeuristicTest, EdgeCases_100)
-// {
-//   auto& rm = umpire::ResourceManager::getInstance();
+TEST(HeuristicTest, EdgeCases_100)
+{
+  auto& rm = umpire::ResourceManager::getInstance();
 
-//   auto h_fun = umpire::strategy::heuristic_percent_releasable(100);
+  auto h_fun = umpire::strategy::heuristic_percent_releasable(100);
 
-//   auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
-//       "host_dyn_pool_h_100", rm.getAllocator("HOST"),
-//       initial_min_size, subsequent_min_size, alignment, h_fun);
+  auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
+      "host_dyn_pool_h_100", rm.getAllocator("HOST"),
+      initial_min_size, subsequent_min_size, alignment, h_fun);
 
-//   auto strategy = alloc.getAllocationStrategy();
-//   auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
+  auto strategy = alloc.getAllocationStrategy();
+  auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
 
-//   if (tracker) {
-//     strategy = tracker->getAllocationStrategy();
-//   }
+  if (tracker) {
+    strategy = tracker->getAllocationStrategy();
+  }
 
-//   auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
+  auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
 
-//   ASSERT_NE(dynamic_pool, nullptr);
+  ASSERT_NE(dynamic_pool, nullptr);
 
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc1 = nullptr;
-//   ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
-//   ASSERT_EQ(alloc.getActualSize(), initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc1 = nullptr;
+  ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
+  ASSERT_EQ(alloc.getActualSize(), initial_min_size);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
-//   ASSERT_EQ(alloc.getActualSize(), initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
+  ASSERT_EQ(alloc.getActualSize(), initial_min_size);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc2 = nullptr;
-//   ASSERT_NO_THROW({ alloc2 = alloc.allocate(initial_min_size); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
-//   ASSERT_EQ(alloc.getActualSize(), 2*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc2 = nullptr;
+  ASSERT_NO_THROW({ alloc2 = alloc.allocate(initial_min_size); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
+  ASSERT_EQ(alloc.getActualSize(), 2*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc3 = nullptr;
-//   ASSERT_NO_THROW({ alloc3 = alloc.allocate(initial_min_size); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc3 = nullptr;
+  ASSERT_NO_THROW({ alloc3 = alloc.allocate(initial_min_size); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc2); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_NO_THROW({ alloc.deallocate(alloc2); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc3); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 2*initial_min_size);
+  ASSERT_NO_THROW({ alloc.deallocate(alloc3); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 2*initial_min_size);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
-// }
+  ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+}
 
-// TEST(HeuristicTest, EdgeCases_0)
-// {
-//   auto& rm = umpire::ResourceManager::getInstance();
+TEST(HeuristicTest, EdgeCases_0)
+{
+  auto& rm = umpire::ResourceManager::getInstance();
 
-//   auto h_fun = umpire::strategy::heuristic_percent_releasable(0);
+  auto h_fun = umpire::strategy::heuristic_percent_releasable(0);
 
-//   auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
-//       "host_dyn_pool_h_0", rm.getAllocator("HOST"),
-//       initial_min_size, subsequent_min_size, alignment, h_fun);
+  auto alloc = rm.makeAllocator<umpire::strategy::DynamicPool>(
+      "host_dyn_pool_h_0", rm.getAllocator("HOST"),
+      initial_min_size, subsequent_min_size, alignment, h_fun);
 
-//   auto strategy = alloc.getAllocationStrategy();
-//   auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
+  auto strategy = alloc.getAllocationStrategy();
+  auto tracker = dynamic_cast<umpire::strategy::AllocationTracker*>(strategy);
 
-//   if (tracker) {
-//     strategy = tracker->getAllocationStrategy();
-//   }
+  if (tracker) {
+    strategy = tracker->getAllocationStrategy();
+  }
 
-//   auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
+  auto dynamic_pool = dynamic_cast<umpire::strategy::DynamicPool*>(strategy);
 
-//   ASSERT_NE(dynamic_pool, nullptr);
+  ASSERT_NE(dynamic_pool, nullptr);
 
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 0);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc1 = nullptr;
-//   ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
-//   ASSERT_EQ(alloc.getActualSize(), initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc1 = nullptr;
+  ASSERT_NO_THROW({ alloc1 = alloc.allocate(16); });
+  ASSERT_EQ(alloc.getActualSize(), initial_min_size);
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 1);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc2 = nullptr;
-//   ASSERT_NO_THROW({ alloc2 = alloc.allocate(initial_min_size); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
-//   ASSERT_EQ(alloc.getActualSize(), 2*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc2 = nullptr;
+  ASSERT_NO_THROW({ alloc2 = alloc.allocate(initial_min_size); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 2);
+  ASSERT_EQ(alloc.getActualSize(), 2*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   void* alloc3 = nullptr;
-//   ASSERT_NO_THROW({ alloc3 = alloc.allocate(initial_min_size); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  void* alloc3 = nullptr;
+  ASSERT_NO_THROW({ alloc3 = alloc.allocate(initial_min_size); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc3); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
+  ASSERT_NO_THROW({ alloc.deallocate(alloc3); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 0);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc2); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 2*initial_min_size);
+  ASSERT_NO_THROW({ alloc.deallocate(alloc2); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 2*initial_min_size);
 
-//   ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
-//   ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
-//   ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
-//   ASSERT_EQ(dynamic_pool->getReleasableSize(), 3*initial_min_size);
-// }
+  ASSERT_NO_THROW({ alloc.deallocate(alloc1); });
+  ASSERT_EQ(dynamic_pool->getBlocksInPool(), 3);
+  ASSERT_EQ(alloc.getActualSize(), 3*initial_min_size);
+  ASSERT_EQ(dynamic_pool->getReleasableSize(), 3*initial_min_size);
+}
 
 #if defined(UMPIRE_ENABLE_NUMA)
 TEST(NumaPolicyTest, EdgeCases) {